bugfix in Math.java
there was a bug in c(int[]... x) at line 930 (and other similar).  x.length was the # arguments to the function, not the length of one of the vectors.  Fixed.  Also, the c(String[]... x) can take a generic type and use .toString() to collect the results.  If already a String, it's a no-op.